+++ lib/src/models/analytics/analytics.dart
@@ -1,5 +1,4 @@
 export 'analytics_event_payload.dart';
 export 'analytics_parameter_keys.dart';
 export 'dashboard/dashboard.dart';
-export 'payloads.dart';
 export 'payloads/payloads.dart';