+++ lib/src/models/config/push_notification_config.g.dart
@@ -47,5 +47,4 @@ const _$PushNotificationProvidersEnumMap = {
 
 const _$PushNotificationSubscriptionDeliveryTypeEnumMap = {
   PushNotificationSubscriptionDeliveryType.breakingOnly: 'breakingOnly',
-  PushNotificationSubscriptionDeliveryType.dailyDigest: 'dailyDigest',
 };