+++ lib/src/models/config/saved_filter_limits.g.dart
@@ -36,5 +36,4 @@ Map<String, dynamic> _$SavedFilterLimitsToJson(SavedFilterLimits instance) =>
 
 const _$PushNotificationSubscriptionDeliveryTypeEnumMap = {
   PushNotificationSubscriptionDeliveryType.breakingOnly: 'breakingOnly',
-  PushNotificationSubscriptionDeliveryType.dailyDigest: 'dailyDigest',
 };