+++ lib/src/models/user_preferences/saved_headline_filter.g.dart
@@ -47,5 +47,4 @@ Map<String, dynamic> _$SavedHeadlineFilterToJson(
 
 const _$PushNotificationSubscriptionDeliveryTypeEnumMap = {
   PushNotificationSubscriptionDeliveryType.breakingOnly: 'breakingOnly',
-  PushNotificationSubscriptionDeliveryType.dailyDigest: 'dailyDigest',
 };