+++ lib/src/models/user_preferences/user_preferences.dart
@@ -1,5 +1,3 @@
 export 'headline_filter_criteria.dart';
 export 'saved_headline_filter.dart';
-export 'saved_source_filter.dart';
-export 'source_filter_criteria.dart';
 export 'user_content_preferences.dart';