+++ test/src/enums/push_notification_subscription_delivery_type_test.dart
@@ -7,12 +7,11 @@ void main() {
       // Arrange: The expected set of enum values.
       const expectedValues = {
         PushNotificationSubscriptionDeliveryType.breakingOnly,
-        PushNotificationSubscriptionDeliveryType.dailyDigest,
       };
 
       // Assert
       // Check that the number of enum values is correct.
-      expect(PushNotificationSubscriptionDeliveryType.values.length, 2);
+      expect(PushNotificationSubscriptionDeliveryType.values.length, 1);
 
       // Check that all expected values are present in the enum's values list.
       for (final value in expectedValues) {