+++ test/src/models/analytics/payloads/headline_filter_updated_payload_test.dart
@@ -14,9 +14,7 @@ void main() {
       filterId: 'filter-123',
       newName: 'New Filter Name',
       pinStatusChangedTo: true,
-      deliveryTypesAdded: [
-        PushNotificationSubscriptionDeliveryType.dailyDigest,
-      ],
+      deliveryTypesAdded: [],
       deliveryTypesRemoved: [
         PushNotificationSubscriptionDeliveryType.breakingOnly,
       ],
@@ -39,7 +37,7 @@ void main() {
           AnalyticsParameterKeys.filterId: 'filter-123',
           'newName': 'New Filter Name',
           'pinStatusChangedTo': true,
-          'deliveryTypesAdded': ['dailyDigest'],
+          'deliveryTypesAdded': <String>[],
           'deliveryTypesRemoved': ['breakingOnly'],
           'newCriteriaSummary': criteriaSummary.toJson(),
         });
@@ -65,9 +63,7 @@ void main() {
           filterId: 'filter-123',
           newName: 'New Filter Name',
           pinStatusChangedTo: true,
-          deliveryTypesAdded: [
-            PushNotificationSubscriptionDeliveryType.dailyDigest,
-          ],
+          deliveryTypesAdded: [],
           deliveryTypesRemoved: [
             PushNotificationSubscriptionDeliveryType.breakingOnly,
           ],