+++ test/src/models/config/user_limits_config_test.dart
@@ -22,7 +22,6 @@ void main() {
           userLimitsConfigFixture.followedItems,
           userLimitsConfigFixture.savedHeadlines,
           userLimitsConfigFixture.savedHeadlineFilters,
-          userLimitsConfigFixture.savedSourceFilters,
           userLimitsConfigFixture.reactionsPerDay,
           userLimitsConfigFixture.commentsPerDay,
           userLimitsConfigFixture.reportsPerDay,